Fix some image references found as part of disconnected test
I updated the dev rpa and prod rpa which I verified with local
dev testing and checking the catalog.  There do seem to be prometheus
images missing from stage and prod and I think stage may have a lot of
rhel9 suffixes that might not be needed?  I don't know for sure about
those so did not make additional changes.

Refs:
 - https://issues.redhat.com/browse/ACM-21584

Signed-off-by: Gus Parvin <gparvin@redhat.com>